Enum reedline::ReedlineEvent

pub enum ReedlineEvent {

    None,
    HistoryHintComplete,
    HistoryHintWordComplete,
    CtrlD,
    CtrlC,
    ClearScreen,
    ClearScrollback,
    Enter,
    Esc,
    Mouse,
    Resize(u16, u16),
    Edit(Vec<EditCommand>),
    Repaint,
    PreviousHistory,
    Up,
    Down,
    Right,
    Left,
    NextHistory,
    SearchHistory,
    Multiple(Vec<ReedlineEvent>),
    UntilFound(Vec<ReedlineEvent>),
    Menu(String),
    MenuNext,
    MenuPrevious,
    MenuUp,
    MenuDown,
    MenuLeft,
    MenuRight,
    MenuPageNext,
    MenuPagePrevious,
    ExecuteHostCommand(String),
    OpenEditor,
    RecordToTill,

}

Reedline supported actions.

Variants

None

No op event

HistoryHintComplete

Complete history hint (default in full)

HistoryHintWordComplete

Complete a single token/word of the history hint

CtrlD

Handle EndOfLine event

Expected Behavior:

On empty line breaks execution to exit with Signal::CtrlD

Secondary behavior EditCommand::Delete

CtrlC

Handle SIGTERM key input

Expected behavior:

Abort entry Run EditCommand::Clear Clear the current undo Bubble up Signal::CtrlC

ClearScreen

Clears the screen and sets prompt to first line

ClearScrollback

Clears the screen and the scrollback buffer

Sets the prompt back to the first line

Enter

Handle enter event

Esc

Esc event

Mouse

Mouse

Resize(u16, u16)

trigger termimal resize

Edit(Vec<EditCommand>)

Run these commands in the editor

Repaint

Trigger full repaint

PreviousHistory

Navigate to the previous historic buffer

Up

Move up to the previous line, if multiline, or up into the historic buffers

Down

Move down to the next line, if multiline, or down through the historic buffers

Right

Move right to the next column, completion entry, or complete hint

Left

Move left to the next column, or completion entry

NextHistory

Navigate to the next historic buffer

SearchHistory

Search the history for a string

Multiple(Vec<ReedlineEvent>)

In vi mode multiple reedline events can be chained while parsing the command or movement characters

UntilFound(Vec<ReedlineEvent>)

Test

Menu(String)

Trigger a menu event. It activates a menu with the event name

MenuNext

Next element in the menu

MenuPrevious

Previous element in the menu

MenuUp

Moves up in the menu

MenuDown

Moves down in the menu

MenuLeft

Moves left in the menu

MenuRight

Moves right in the menu

MenuPageNext

Move to the next history page

MenuPagePrevious

Move to the previous history page

ExecuteHostCommand(String)

Way to bind the execution of a whole command (directly returning from crate::Reedline::read_line()) to a keybinding

OpenEditor

Open text editor

RecordToTill

Record vi to or till motion