Module reedline::menu_functions

Collection of common functions that can be used to create menus

Structs

ParseResult

Index result obtained from parsing a string with an index marker For example, the next string: “this is an example :10”

Enums

ParseAction

Direction of the index found in the string

Functions

find_common_string

Finds index for the common string in a list of suggestions

parse_selection_char

Splits a string that contains a marker character

string_difference

Finds different string between two strings