Struct reedline::ColumnarMenu

pub struct ColumnarMenu { /* private fields */ }

Menu to present suggestions in a columnar fashion It presents a description of the suggestion if available

Implementations

impl ColumnarMenu

pub fn with_name(self, name: &str) -> Self

Menu builder with new name

pub fn with_text_style(self, text_style: Style) -> Self

Menu builder with new value for text style

pub fn with_selected_text_style(self, selected_text_style: Style) -> Self

Menu builder with new value for text style

pub fn with_description_text_style(self, description_text_style: Style) -> Self

Menu builder with new value for text style

pub fn with_columns(self, columns: u16) -> Self

Menu builder with new columns value

pub fn with_column_width(self, col_width: Option<usize>) -> Self

Menu builder with new column width value

pub fn with_column_padding(self, col_padding: usize) -> Self

Menu builder with new column width value

pub fn with_marker(self, marker: String) -> Self

Menu builder with marker

pub fn with_only_buffer_difference(self, only_buffer_difference: bool) -> Self

Menu builder with new only buffer difference