reedline
Struct FileBackedHistory

pub struct FileBackedHistory { /* private fields */ }

Stateful history that allows up/down-arrow browsing with an internal cursor.

Can optionally be associated with a newline separated history file using the FileBackedHistory::with_file() constructor. Similar to bash’s behavior without HISTTIMEFORMAT. (See https://www.gnu.org/software/bash/manual/html_node/Bash-History-Facilities.html) If the history is associated to a file all new changes within a given history capacity will be written to disk when History is dropped.

Implementations

impl FileBackedHistory

pub fn new(capacity: usize) -> Result<Self>

Creates a new in-memory history that remembers n <= capacity elements

pub fn with_file(capacity: usize, file: PathBuf) -> Result<Self>

Creates a new history with an associated history file.

History file format: commands separated by new lines. If file exists file will be read otherwise empty file will be created.

Side effects: creates all nested directories to the file