reedline
Struct LineBuffer

pub struct LineBuffer { /* private fields */ }

In memory representation of the entered line(s) including a cursor position to facilitate cursor based editing.

Implementations

impl LineBuffer

pub fn new() -> LineBuffer

Create a line buffer instance

pub fn is_empty(&self) -> bool

Check to see if the line buffer is empty

pub fn is_valid(&self) -> bool

Check if the line buffer is valid utf-8 and the cursor sits on a valid grapheme boundary

pub fn insertion_point(&self) -> usize

Gets the current edit position

pub fn set_insertion_point(&mut self, offset: usize)

Sets the current edit position

Unicode safety:

Not checked, improper use may cause panics in following operations

pub fn get_buffer(&self) -> &str

Output the current line in the multiline buffer

pub fn set_buffer(&mut self, buffer: String)

Set to a single line of buffer and reset the InsertionPoint cursor to the end

pub fn line(&self) -> usize

Calculates the current the user is on

Zero-based index

pub fn num_lines(&self) -> usize

Counts the number of lines in the buffer

pub fn ends_with(&self, c: char) -> bool

Checks to see if the buffer ends with a given character

pub fn move_to_start(&mut self)

Reset the insertion point to the start of the buffer

pub fn move_to_line_start(&mut self)

Move the cursor before the first character of the line

pub fn move_to_line_end(&mut self)

Move cursor position to the end of the line

Insertion will append to the line. Cursor on top of the potential \n or \r of \r\n

pub fn move_to_end(&mut self)

Set the insertion point behind the last character.

pub fn len(&self) -> usize

Get the length of the buffer

pub fn find_current_line_end(&self) -> usize

Returns where the current line terminates

Either:

end of buffer (len())

\n or \r\n (on the first byte)

pub fn grapheme_right_index(&self) -> usize

Cursor position behind the next unicode grapheme to the right

pub fn grapheme_left_index(&self) -> usize

Cursor position in front of the next unicode grapheme to the left

pub fn word_right_index(&self) -> usize

Cursor position behind the next word to the right

pub fn big_word_right_index(&self) -> usize

Cursor position behind the next WORD to the right

pub fn word_right_end_index(&self) -> usize

Cursor position at end of the next word to the right

pub fn big_word_right_end_index(&self) -> usize

Cursor position at end of the next WORD to the right

pub fn word_right_start_index(&self) -> usize

Cursor position in front of the next word to the right

pub fn big_word_right_start_index(&self) -> usize

Cursor position in front of the next WORD to the right

pub fn word_left_index(&self) -> usize

Cursor position in front of the next word to the left

pub fn big_word_left_index(&self) -> usize

Cursor position in front of the next WORD to the left

pub fn next_whitespace(&self) -> usize

Cursor position on the next whitespace

pub fn move_right(&mut self)

Move cursor position behind the next unicode grapheme to the right

pub fn move_left(&mut self)

Move cursor position in front of the next unicode grapheme to the left

pub fn move_word_left(&mut self)

Move cursor position in front of the next word to the left

pub fn move_big_word_left(&mut self)

Move cursor position in front of the next WORD to the left

pub fn move_word_right(&mut self)

Move cursor position behind the next word to the right

pub fn move_word_right_start(&mut self)

Move cursor position to the start of the next word

pub fn move_big_word_right_start(&mut self)

Move cursor position to the start of the next WORD

pub fn move_word_right_end(&mut self)

Move cursor position to the end of the next word

pub fn move_big_word_right_end(&mut self)

Move cursor position to the end of the next WORD

pub fn insert_char(&mut self, c: char)

Insert a single character at the insertion point and move right

pub fn insert_str(&mut self, string: &str)

Insert &str at the cursor position in the current line.

Sets cursor to end of inserted string

Unicode safety:

Does not validate the incoming string or the current cursor position

pub fn insert_newline(&mut self)

Inserts the system specific new line character

On Unix systems LF ("\n")

On Windows CRLF ("\r\n")

pub fn clear(&mut self)

Empty buffer and reset cursor

pub fn clear_to_end(&mut self)

Clear everything beginning at the cursor to the right/end. Keeps the cursor at the end.

pub fn clear_to_line_end(&mut self)

Clear beginning at the cursor up to the end of the line. Newline character at the end remains.

pub fn clear_to_insertion_point(&mut self)

Clear from the start of the buffer to the cursor. Keeps the cursor at the beginning of the line/buffer.

pub fn clear_range_safe(&mut self, start: usize, end: usize)

Clear all contents between start and end and change insertion point if necessary.

If the cursor is located between start and end it is adjusted to start. If the cursor is located after end it is adjusted to stay at its current char boundary.

pub fn replace_range<R>(&mut self, range: R, replace_with: &str)
where R: RangeBounds<usize>,

Substitute text covered by range in the current line

Safety: Does not change the insertion point/offset and is thus not unicode safe!

pub fn on_whitespace(&self) -> bool

Checks to see if the current edit position is pointing to whitespace

pub fn grapheme_right(&self) -> &str

Get the grapheme immediately to the right of the cursor, if any

pub fn grapheme_left(&self) -> &str

Get the grapheme immediately to the left of the cursor, if any

pub fn current_word_range(&self) -> Range<usize>

Gets the range of the word the current edit position is pointing to

pub fn current_line_range(&self) -> Range<usize>

Range over the current line

Starts on the first non-newline character and is an exclusive range extending beyond the potential carriage return and line feed characters terminating the line

pub fn uppercase_word(&mut self)

Uppercases the current word

pub fn lowercase_word(&mut self)

Lowercases the current word

pub fn switchcase_char(&mut self)

Switches the ASCII case of the current char

pub fn capitalize_char(&mut self)

Capitalize the character at insertion point (or the first character following the whitespace at the insertion point) and move the insertion point right one grapheme.

pub fn delete_left_grapheme(&mut self)

Deletes on grapheme to the left

pub fn delete_right_grapheme(&mut self)

Deletes one grapheme to the right

pub fn delete_word_left(&mut self)

Deletes one word to the left

pub fn delete_word_right(&mut self)

Deletes one word to the right

pub fn swap_words(&mut self)

Swaps current word with word on right

pub fn swap_graphemes(&mut self)

Swaps current grapheme with grapheme on right

pub fn move_line_up(&mut self)

Moves one line up

pub fn move_line_down(&mut self)

Moves one line down

pub fn is_cursor_at_first_line(&self) -> bool

Checks to see if the cursor is on the first line of the buffer

pub fn is_cursor_at_last_line(&self) -> bool

Checks to see if the cursor is on the last line of the buffer

pub fn find_char_right(&self, c: char, current_line: bool) -> Option<usize>

Finds index for the first occurrence of a char to the right of offset

pub fn find_char_left(&self, c: char, current_line: bool) -> Option<usize>

Finds index for the first occurrence of a char to the left of offset

pub fn move_right_until(&mut self, c: char, current_line: bool) -> usize

Moves the insertion point until the next char to the right

pub fn move_right_before(&mut self, c: char, current_line: bool) -> usize

Moves the insertion point before the next char to the right

pub fn move_left_until(&mut self, c: char, current_line: bool) -> usize

Moves the insertion point until the next char to the left of offset

pub fn move_left_before(&mut self, c: char, current_line: bool) -> usize

Moves the insertion point before the next char to the left of offset

pub fn delete_right_until_char(&mut self, c: char, current_line: bool)

Deletes until first character to the right of offset

pub fn delete_right_before_char(&mut self, c: char, current_line: bool)

Deletes before first character to the right of offset

pub fn delete_left_until_char(&mut self, c: char, current_line: bool)

Deletes until first character to the left of offset

pub fn delete_left_before_char(&mut self, c: char, current_line: bool)

Deletes before first character to the left of offset