Struct reedline::StyledText

pub struct StyledText {
    pub buffer: Vec<(Style, String)>,
}

A representation of a buffer with styling, used for doing syntax highlighting

Fields
buffer: Vec<(Style, String)>

The component, styled parts of the text

Implementations

impl StyledText

pub const fn new() -> Self

Construct a new StyledText

pub fn push(&mut self, styled_string: (Style, String))

Add a new styled string to the buffer

pub fn render_around_insertion_point(
&self,
insertion_point: usize,
prompt: &dyn Prompt,
use_ansi_coloring: bool
) -> (String, String)

Render the styled string. We use the insertion point to render around so that we can properly write out the styled string to the screen and find the correct place to put the cursor. This assumes a logic that prints the first part of the string, saves the cursor position, prints the second half, and then restores the cursor position

Also inserts the multiline continuation prompt

pub fn render_simple(&self) -> String

Apply the ANSI style formatting to the full string.

pub fn raw_string(&self) -> String

Get the unformatted text as a single continuous string.