Trait reedline::EditMode

pub trait EditMode: Send {
    fn parse_event(&mut self, event: Event) -> ReedlineEvent;
    fn edit_mode(&self) -> PromptEditMode;
}

Define the style of parsing for the edit events Available default options:

Emacs

Vi

Required Methods

fn parse_event(&mut self, event: Event) -> ReedlineEvent

Translate the given user input event into what the LineEditor understands

fn edit_mode(&self) -> PromptEditMode

What to display in the prompt indicator