reedline
Trait Validator

pub trait Validator: Send {
    // Required method
    fn validate(&self, line: &str) -> ValidationResult;
}

The syntax validation trait. Implementers of this trait will check to see if the current input is incomplete and spans multiple lines

Required Methods

fn validate(&self, line: &str) -> ValidationResult

The action that will handle the current buffer as a line and return the corresponding validation