reedline
Enum EditCommand

#[non_exhaustive]
pub enum EditCommand {

    MoveToStart {
        select: bool,
    },
    MoveToLineStart {
        select: bool,
    },
    MoveToEnd {
        select: bool,
    },
    MoveToLineEnd {
        select: bool,
    },
    MoveLeft {
        select: bool,
    },
    MoveRight {
        select: bool,
    },
    MoveWordLeft {
        select: bool,
    },
    MoveBigWordLeft {
        select: bool,
    },
    MoveWordRight {
        select: bool,
    },
    MoveWordRightStart {
        select: bool,
    },
    MoveBigWordRightStart {
        select: bool,
    },
    MoveWordRightEnd {
        select: bool,
    },
    MoveBigWordRightEnd {
        select: bool,
    },
    MoveToPosition {
        position: usize,
        select: bool,
    },
    InsertChar(char),
    InsertString(String),
    InsertNewline,
    ReplaceChar(char),
    ReplaceChars(usize, String),
    Backspace,
    Delete,
    CutChar,
    BackspaceWord,
    DeleteWord,
    Clear,
    ClearToLineEnd,
    Complete,
    CutCurrentLine,
    CutFromStart,
    CutFromLineStart,
    CutToEnd,
    CutToLineEnd,
    CutWordLeft,
    CutBigWordLeft,
    CutWordRight,
    CutBigWordRight,
    CutWordRightToNext,
    CutBigWordRightToNext,
    PasteCutBufferBefore,
    PasteCutBufferAfter,
    UppercaseWord,
    LowercaseWord,
    CapitalizeChar,
    SwitchcaseChar,
    SwapWords,
    SwapGraphemes,
    Undo,
    Redo,
    CutRightUntil(char),
    CutRightBefore(char),
    MoveRightUntil {
        c: char,
        select: bool,
    },
    MoveRightBefore {
        c: char,
        select: bool,
    },
    CutLeftUntil(char),
    CutLeftBefore(char),
    MoveLeftUntil {
        c: char,
        select: bool,
    },
    MoveLeftBefore {
        c: char,
        select: bool,
    },
    SelectAll,
    CutSelection,
    CopySelection,
    Paste,

}

Editing actions which can be mapped to key bindings.

Executed by Reedline::run_edit_commands()

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

MoveToStart

Move to the start of the buffer

Fields

select: bool

Select the text between the current cursor position and destination

MoveToLineStart

Move to the start of the current line

Fields

select: bool

Select the text between the current cursor position and destination

MoveToEnd

Move to the end of the buffer

Fields

select: bool

Select the text between the current cursor position and destination

MoveToLineEnd

Move to the end of the current line

Fields

select: bool

Select the text between the current cursor position and destination

MoveLeft

Move one character to the left

Fields

select: bool

Select the text between the current cursor position and destination

MoveRight

Move one character to the right

Fields

select: bool

Select the text between the current cursor position and destination

MoveWordLeft

Move one word to the left

Fields

select: bool

Select the text between the current cursor position and destination

MoveBigWordLeft

Move one WORD to the left

Fields

select: bool

Select the text between the current cursor position and destination

MoveWordRight

Move one word to the right

Fields

select: bool

Select the text between the current cursor position and destination

MoveWordRightStart

Move one word to the right, stop at start of word

Fields

select: bool

Select the text between the current cursor position and destination

MoveBigWordRightStart

Move one WORD to the right, stop at start of WORD

Fields

select: bool

Select the text between the current cursor position and destination

MoveWordRightEnd

Move one word to the right, stop at end of word

Fields

select: bool

Select the text between the current cursor position and destination

MoveBigWordRightEnd

Move one WORD to the right, stop at end of WORD

Fields

select: bool

Select the text between the current cursor position and destination

MoveToPosition

Move to position

Fields

position: usize

Position to move to

select: bool

Select the text between the current cursor position and destination

InsertChar(char)

Insert a character at the current insertion point

InsertString(String)

Insert a string at the current insertion point

InsertNewline

Inserts the system specific new line character

On Unix systems LF ("\n")

On Windows CRLF ("\r\n")

ReplaceChar(char)

Replace a character

ReplaceChars(usize, String)

Replace characters with string

Backspace

Backspace delete from the current insertion point

Delete

Delete in-place from the current insertion point

CutChar

Cut the grapheme right from the current insertion point

BackspaceWord

Backspace delete a word from the current insertion point

DeleteWord

Delete in-place a word from the current insertion point

Clear

Clear the current buffer

ClearToLineEnd

Clear to the end of the current line

Complete

Insert completion: entire completion if there is only one possibility, or else up to shared prefix.

CutCurrentLine

Cut the current line

CutFromStart

Cut from the start of the buffer to the insertion point

CutFromLineStart

Cut from the start of the current line to the insertion point

CutToEnd

Cut from the insertion point to the end of the buffer

CutToLineEnd

Cut from the insertion point to the end of the current line

CutWordLeft

Cut the word left of the insertion point

CutBigWordLeft

Cut the WORD left of the insertion point

CutWordRight

Cut the word right of the insertion point

CutBigWordRight

Cut the word right of the insertion point

CutWordRightToNext

Cut the word right of the insertion point and any following space

CutBigWordRightToNext

Cut the WORD right of the insertion point and any following space

PasteCutBufferBefore

Paste the cut buffer in front of the insertion point (Emacs, vi P)

PasteCutBufferAfter

Paste the cut buffer in front of the insertion point (vi p)

UppercaseWord

Upper case the current word

LowercaseWord

Lower case the current word

CapitalizeChar

Capitalize the current character

SwitchcaseChar

Switch the case of the current character

SwapWords

Swap the current word with the word to the right

SwapGraphemes

Swap the current grapheme/character with the one to the right

Undo

Undo the previous edit command

Redo

Redo an edit command from the undo history

CutRightUntil(char)

CutUntil right until char

CutRightBefore(char)

CutUntil right before char

MoveRightUntil

CutUntil right until char

Fields

c: char

Char to move towards

select: bool

Select the text between the current cursor position and destination

MoveRightBefore

CutUntil right before char

Fields

c: char

Char to move towards

select: bool

Select the text between the current cursor position and destination

CutLeftUntil(char)

CutUntil left until char

CutLeftBefore(char)

CutUntil left before char

MoveLeftUntil

Move left until char

Fields

c: char

Char to move towards

select: bool

Select the text between the current cursor position and destination

MoveLeftBefore

Move left before char

Fields

c: char

Char to move towards

select: bool

Select the text between the current cursor position and destination

SelectAll

Select whole input buffer

CutSelection

Cut selection to local buffer

CopySelection

Copy selection to local buffer

Paste

Paste content from local buffer at the current cursor position

Implementations

impl EditCommand

pub fn edit_type(&self) -> EditType

Determine if a certain operation should be undoable or if the operations should be coalesced for undoing