reedline
Struct ColumnarMenu

pub struct ColumnarMenu { /* private fields */ }

Menu to present suggestions in a columnar fashion It presents a description of the suggestion if available

Implementations

impl ColumnarMenu

pub fn with_columns(self, columns: u16) -> Self

Menu builder with new columns value

pub fn with_column_width(self, col_width: Option<usize>) -> Self

Menu builder with new column width value

pub fn with_column_padding(self, col_padding: usize) -> Self

Menu builder with new column width value