reedline
Trait Menu

pub trait Menu: Send {

    // Required methods
    fn is_active(&self) -> bool;
    fn menu_event(&mut self, event: MenuEvent);
    fn can_quick_complete(&self) -> bool;
    fn can_partially_complete(
        &mut self,
        values_updated: bool,
        editor: &mut Editor,
        completer: &mut dyn Completer,
    ) -> bool;
    fn update_values(
        &mut self,
        editor: &mut Editor,
        completer: &mut dyn Completer,
    );
    fn update_working_details(
        &mut self,
        editor: &mut Editor,
        completer: &mut dyn Completer,
        painter: &Painter,
    );
    fn replace_in_buffer(&self, editor: &mut Editor);
    fn menu_required_lines(&self, terminal_columns: u16) -> u16;
    fn menu_string(
        &self,
        available_lines: u16,
        use_ansi_coloring: bool,
    ) -> String;
    fn min_rows(&self) -> u16;
    fn get_values(&self) -> &[Suggestion];

    // Provided methods
    fn settings(&self) -> &MenuSettings { ... }
    fn name(&self) -> &str { ... }
    fn indicator(&self) -> &str { ... }
    fn set_cursor_pos(&mut self, _pos: (u16, u16)) { ... }

}

Trait that defines how a menu will be printed by the painter

Required Methods

fn is_active(&self) -> bool

Checks if the menu is active

fn menu_event(&mut self, event: MenuEvent)

Selects what type of event happened with the menu

fn can_quick_complete(&self) -> bool

A menu may not be allowed to quick complete because it needs to stay active even with one element

fn can_partially_complete( &mut self, values_updated: bool, editor: &mut Editor, completer: &mut dyn Completer, ) -> bool

The completion menu can try to find the common string and replace it in the given line buffer

fn update_values(&mut self, editor: &mut Editor, completer: &mut dyn Completer)

Updates the values presented in the menu This function needs to be defined in the trait because when the menu is activated or the quick_completion option is true, the len of the values is calculated to know if there is only one value so it can be selected immediately

fn update_working_details( &mut self, editor: &mut Editor, completer: &mut dyn Completer, painter: &Painter, )

The working details of a menu are values that could change based on the menu conditions before it being printed, such as the number or size of columns, etc. In this function should be defined how the menu event is treated since it is called just before painting the menu

fn replace_in_buffer(&self, editor: &mut Editor)

Indicates how to replace in the line buffer the selected value from the menu

fn menu_required_lines(&self, terminal_columns: u16) -> u16

Calculates the real required lines for the menu considering how many lines wrap the terminal or if entries have multiple lines

fn menu_string(&self, available_lines: u16, use_ansi_coloring: bool) -> String

Creates the menu representation as a string which will be painted by the painter

fn min_rows(&self) -> u16

Minimum rows that should be displayed by the menu

fn get_values(&self) -> &[Suggestion]

Gets cached values from menu that will be displayed

Provided Methods

fn settings(&self) -> &MenuSettings

Get MenuSettings

fn name(&self) -> &str

Menu name

fn indicator(&self) -> &str

Menu indicator

fn set_cursor_pos(&mut self, _pos: (u16, u16))

Sets the position of the cursor (currently only required by the IDE menu)